Added sliding movegen.
